# docs/apis/api.yaml
openapi: 3.0.3
...
info:
  title: Jiku api — contrato HTTP
  version: 1.0.0
  description: |
    Única puerta HTTP del producto. Autentica contra Zitadel, autoriza por rol y por entidad,
    **lee la base de datos directamente** y convierte toda mutación en un comando publicado en
    NATS que atiende `core`.

    **Las escrituras pasan por el bus.** Los endpoints marcados con `x-bus-command` publican un
    comando y esperan respuesta. Eso agrega **tres** modos de falla a todos ellos:

    - **`503` (`service_unavailable`) — problema de despliegue.** No hay **ningún suscriptor** del
      subject. El server contesta *"no responders"* **en milisegundos**, no al vencer el timeout.
      La operación **no ocurrió**, así que reintentar es seguro.
    - **`504` (`gateway_timeout`) — problema de performance.** Había alguien escuchando, pero la
      respuesta no llegó dentro de `NATS_REQUEST_TIMEOUT_MS` (default 5000 ms). **La operación
      pudo haber ocurrido**: sin JetStream no hay acuse, así que reintentar a ciegas puede
      duplicar.
    - **Los códigos de error que devuelve core**, traducidos a HTTP por `httpStatusFor`.

    Los dos primeros los genera la api en su `catch`; **`gateway_timeout` NO es un `errorCode` de
    core** y por eso **no está en `STATUS_BY_ERROR_CODE`**. Separarlos es lo que permite distinguir
    "no desplegué core" de "core está lento" sin mirar los logs. El contrato del bus está en
    [core.yaml](./core.yaml).

...

...

...

...
    >
...

...

...
    produce hoy.
  license:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /opus/projects/{projid}/requirements:
    get:
      tags: [opus]
      summary: Requisitos de un proyecto en el portal
      description: |
        **Devuelve solo requisitos con `visibilityLevel: 'public'`**, para todo rol —
        `user` y `admin` incluidos. El recorte es de la superficie, no del caller: el portal es
        la pantalla que se comparte con el cliente. No se puede desactivar por query.

        **Nota sobre el código:** los filtros se validan con `validateBodyFields`, no con
        `validateQueryParams` (`opus-projects-projid-requirements-get.ts:63`), así que el esquema
        se aplica al cuerpo de una request GET. Los parámetros que el handler realmente lee son
        los de query.
      x-roles: [user, external-user]
      parameters:
        - $ref: '#/components/parameters/ProjIdPath'
        - name: state
          in: query
          description: Uno o varios estados
          schema:
            oneOf:
              - { $ref: '#/components/schemas/RequirementState' }
              - type: array
                items: { $ref: '#/components/schemas/RequirementState' }
        - name: sort
          in: query
          schema: { type: string, enum: [createdAt, title], default: createdAt }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 20, default: 20 }
        - name: skip
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/Requirement'
                    - type: object
                      properties:
                        creator: { $ref: '#/components/schemas/AuthorUser' }
        '400': { $ref: '#/components/responses/InvalidFields' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '404': { $ref: '#/components/responses/ProjectNotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /opus/requirements/{reqid}:
    parameters:
      - $ref: '#/components/parameters/ReqIdPath'
    get:
      tags: [opus]
      summary: Detalle de un requisito en el portal
      description: |
        **Responde 404 `requirement_not_found` si el requisito es `internal`**, para todo rol.
        El 404 es idéntico al de un id inexistente, a propósito: distinguirlos confirmaría que
        el requisito existe (mismo criterio que el recorte externo de `core`).

        Filtra las actividades por visibilidad: un `external-user` no ve las `internal`.

        **La respuesta real NO es `RequirementWithRelations`.** Este handler es el único de la
        api que no serializa el modelo: `sendResponse` construye el objeto campo por campo
        (`opus-requirements-id-get.ts:61-96`). Lo que devuelve, con los nombres exactos:

        - `creator` — un `AuthorUser` (el `pie-autoria` de `opus-web`).
        - `requirementActivity[]` — el feed y los comentarios; cada entrada trae `user`, también
          un `AuthorUser`.
        - `subscriptors[]` — **tres campos** (`id`, `name`, `email`) y nada más: es un
          **selector** de suscriptores, no una autoría, así que **no** lleva `identityType`.
        - `lastActivityAt` — derivado de la última actividad pública.

        El `$ref` de abajo quedó desalineado con el handler **antes** de S-019 y sigue así: su
        arreglo va por `/product-change-technical-definition`.
      x-roles: [user, external-user]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RequirementWithRelations' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '404': { $ref: '#/components/responses/RequirementNotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    patch:
      tags: [opus]
      summary: Cambia estado o prioridad de un requisito
      description: |
        Publica `requirements.{id}.edit` con solo esos dos campos.

        **Responde 404 `requirement_not_found` si el requisito es `internal`**, y corta antes de
        publicar el comando: lo que el portal no muestra, tampoco lo edita.

        **Declara `hasAnyRole(['user', 'admin'])` — sin `external-user`**, a diferencia del resto
        de la superficie opus (`opus-requirements-id-patch.ts:30`). Si es intencional o un
        descuido está sin confirmar.
      x-roles: [user, admin]
      x-bus-command: requirements.{id}.edit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                state: { $ref: '#/components/schemas/RequirementState' }
                priority: { $ref: '#/components/schemas/RequirementPriority' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Acknowledgement' }
        '400': { $ref: '#/components/responses/InvalidFields' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '404': { $ref: '#/components/responses/RequirementNotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/BusUnavailable' }
        '504': { $ref: '#/components/responses/GatewayTimeout' }
  /opus/requirements/{reqid}/comments:
    post:
      tags: [opus]
      summary: Comenta un requisito desde el portal
      description: |
        Publica `requirements.{id}.comment`. **Sin `visibilityLevel`**: un comentario del portal
        no puede ser interno.

        **Responde 404 `requirement_not_found` si el requisito es `internal`**, y corta antes de
        publicar el comando.
      x-roles: [user, external-user]
      x-bus-command: requirements.{id}.comment
      parameters:
        - $ref: '#/components/parameters/ReqIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [comment]
              properties:
                comment: { type: string }
                fileIds:
                  $ref: '#/components/schemas/FileIds'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RequirementActivity' }
        '400': { $ref: '#/components/responses/InvalidFields' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '404': { $ref: '#/components/responses/RequirementNotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/BusUnavailable' }
        '504': { $ref: '#/components/responses/GatewayTimeout' }
  /opus/requirements/{reqid}/subscriptors:
    post:
      tags: [opus]
      summary: Suscribe un usuario a un requisito
      description: |
        Publica `requirements.{id}.subscriptors.new`. Valida que el usuario a suscribir exista y
        tenga permiso sobre el proyecto.

        **Responde 404 `requirement_not_found` si el requisito es `internal`.**
      x-roles: [external-user]
      x-bus-command: requirements.{id}.subscriptors.new
      parameters:
        - $ref: '#/components/parameters/ReqIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId: { type: string, description: Id de usuario de Zitadel }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Acknowledgement' }
        '400':
          description: Validación fallida, o `already_subscribed` desde core
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccessDenied' }
        '404':
          description: El requisito o el usuario no existen
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/BusUnavailable' }
        '504': { $ref: '#/components/responses/GatewayTimeout' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT RS256 emitido por Zitadel, verificado contra su JWKS
        (`{IDENTITY_URL}/oauth/v2/keys`), resincronizado con reintentos cuando aparece un `kid`
        desconocido.

        **Solo por header.** El fallback por query param (`?jwt=`) se eliminó: los query params
        quedan en logs de acceso, historial y `Referer`.

        Los roles salen del claim `urn:zitadel:iam:org:project:roles`, como las claves del objeto.
        Roles del producto: `admin`, `user`, `external-user`.

        > **REQ-007 (S-034) — `- ELIMINADO`.** Acá decía: *"El `sub` tiene que existir en la
        > tabla `users` o la respuesta es 401 `user_not_found`"*. **Ese 401 ya no existe en
        > ninguna de las 61 rutas.**
        >
        > `validateToken` **dejó de hacer `User.findByPk(sub)`** y arma `req.user` con
        > `{ id, name, username, email, roles }` **del claim ya verificado** — que es la fuente
        > que ADR-007 declara autoritativa (*"los roles DEBEN leerse del claim […] NO SE DEBEN
        > almacenar roles en la base ni derivarlos de otra fuente"*). Con eso:
        >
        > - una identidad autenticada **opera aunque no tenga fila previa**, y la fila queda
        >   creada la primera vez que ejecuta un comando (`core` la espeja desde el sobre)
        > - **se elimina un `SELECT` por request** en toda la superficie HTTP
        > - `POST /api/auth/present` **deja de necesitar su excepción**: ya no hay chequeo del
        >   que eximirlo
        >
        > **El `401 unauthorized` (sin token / token inválido) NO cambia.**
  parameters:
...
    ProjIdPath:
      name: projid
      in: path
      required: true
      schema: { type: integer }
    ReqIdPath:
      name: reqid
      in: path
      required: true
      schema: { type: integer }
...
  responses:
    Unauthorized:
      # REQ-007 (S-034): `- ELIMINADO` el ejemplo `userNotFound`. El `sub` ya NO tiene que estar
      # en `users`: `req.user` se arma del claim. Queda una sola causa de 401.
      description: Sin token o token inválido
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            unauthorized:
              value: { code: unauthorized, message: Unauthorized }
    AccessDenied:
      description: Autenticado, pero el rol o el permiso de proyecto no alcanzan
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: access_denied, message: Access denied }
    InvalidFields:
      description: Validación de Joi fallida. Devuelve **solo el primer** error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: invalid_fields, message: 'Invalid field - "title" is required' }
...
    ProjectNotFound:
      description: El proyecto no existe
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: project_not_found, message: Project not found }
    RequirementNotFound:
      description: El requisito no existe
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: requirement_not_found, message: Requirement not found }
    InternalError:
      description: Error inesperado. El detalle va al log, no a la respuesta
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: internal_error, message: Internal error }
    BusUnavailable:
      description: |
        **No hay ningún suscriptor del subject:** el server contesta *"no responders"* **en
        milisegundos**, no al vencer el timeout. Es la señal de un problema de **despliegue**.

        **La operación no ocurrió**, y sin JetStream no hay reintento — así que **reintentar es
        seguro**.

        No cubre el timeout: eso es `504` (`GatewayTimeout`).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            code: service_unavailable
            message: El servicio no está disponible en este momento
    GatewayTimeout:
      description: |
        **La respuesta no llegó dentro de `NATS_REQUEST_TIMEOUT_MS`** (default 5000 ms). Había
        alguien escuchando y no contestó a tiempo: es la señal de un problema de **performance**,
        no de despliegue.

        **LA OPERACIÓN PUDO HABER OCURRIDO.** Es el modo de falla más confuso del sistema: sin
        JetStream no hay acuse ni idempotencia (ADR-002), así que core puede haber escrito y la
        respuesta haberse perdido. **Reintentar a ciegas puede duplicar.** El `message` no invita
        a reintentar, a propósito.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            code: gateway_timeout
            message: La operación tardó demasiado
  schemas:
...
    Acknowledgement:
      type: object
      description: Acuse de una edición que no devuelve el recurso
      properties:
        code: { type: string }
        message: { type: string }

    # ---- enums ----
...
    RequirementState:
      type: string
      description: |
        Los valores salen del enum del modelo compartido
        (`packages/models/src/requirement.model.ts`). `en_cola` se agregó en la migración
        `20260707_01`. Default en la base: `analisis`
      enum: [analisis, planificacion, en_cola, desarrollo, revision, resuelto, cancelado]
    RequirementType:
      type: string
      enum: [funcionalidad, mejora, incidencia, otro]
    RequirementPriority:
      type: string
      enum: [sin_prioridad, baja, media, alta, urgente]
...
    AuthorUser:
      type: object
      description: |
        El usuario cuando aparece como **autor** de algo: creador de un proyecto, un requisito o
        una tarea, y autor de una entrada de actividad o de un comentario.

        Es un schema distinto de `User` a propósito (S-019, CA-1 y CA-2). Declara `identityType`
        porque la interfaz necesita distinguir a una persona de una identidad automática y
        marcarla; **`User` no lo declara** porque es alcanzable por `external-user` a través de
        `GET /opus/projects/{projid}/users`, un **selector de personas** ya filtrado a
        `identityType: 'person'`, donde el campo sería constante.

        **`roles` no sale acá ni en ningún otro schema.** Ese es el argumento de seguridad de
        D-12 de REQ-005, y sigue en pie: exponer el modelo de roles del producto en una respuesta
        HTTP sería filtrar información de autorización. La garantía no la da este schema: la dan
        los `include` acotados de la api.

        **No cubre a todo `User` embebido.** Un `uploader` de adjunto, un suscriptor y el
        `creator` de `GET /projects` **no** son `AuthorUser`: siguen devolviendo tres campos. El
        arreglo completo de la deriva —un `UserSummary` para todo usuario embebido— sigue
        pendiente y va por `/product-change-technical-definition`.
      properties:
        id: { type: string }
        name: { type: string }
        email:
          type: [string, 'null']
          format: email
          description: |
            **`null` para una identidad de servicio**, que es exactamente la que `identityType`
            marca: los dos campos describen la misma distinción desde dos ángulos. Ver `User.email`.
        identityType:
          type: string
          enum: [person, service]
          description: |
            Marca de identidad automática en la interfaz. `service` es una identidad de servicio
            —un conector externo, no una persona—, y el front renderiza la marca **solo** si el
            valor es exactamente `service`. Los dos valores están en inglés porque no los elige
            el producto: son el `type` de `deploy/nats/auth-callout/rules.yaml`. La palabra que
            ve el usuario la elige el front.
...
    Requirement:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        type:
          allOf: [{ $ref: '#/components/schemas/RequirementType' }]
          nullable: true
        priority: { $ref: '#/components/schemas/RequirementPriority' }
        state: { $ref: '#/components/schemas/RequirementState' }
        visibilityLevel: { $ref: '#/components/schemas/VisibilityLevel' }
        estimatedFinishDate: { type: string, format: date, nullable: true }
        projectId: { type: integer }
        createdBy: { type: string }
        resolutionType:
          allOf: [{ $ref: '#/components/schemas/RequirementResolution' }]
          nullable: true
        resolutionConclusion: { type: string, nullable: true }
        resolutionComment: { type: string, nullable: true }
        scope: { type: string, nullable: true }
        technicalSolution: { type: string, nullable: true }
        acceptanceCriteria: { type: string, nullable: true }
        tags:
          type: array
          nullable: true
          items: { $ref: '#/components/schemas/Tag' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    RequirementWithRelations:
      description: |
        **Lo referencian cinco respuestas, y solo dos devuelven `creator` de verdad:**
        `GET /requirements/{reqid}` y `GET /opus/requirements/{reqid}`. `GET /requirements`,
        `POST /requirements` y `PATCH /requirements/{reqid}` **no** tienen el `include` de
        `creator` (ni el del feed), así que para esos tres el schema sobredeclara. La deriva es
        preexistente —Hallazgo 4 del plan de S-015 (api)— y su arreglo va por
        `/product-change-technical-definition`.

        **`GET /opus/requirements/{reqid}` además no devuelve este schema:** su handler arma la
        respuesta campo por campo (proyección a mano). Ver la `description` de ese endpoint.
      allOf:
        - $ref: '#/components/schemas/Requirement'
        - type: object
          properties:
            project: { $ref: '#/components/schemas/ProjectSummary' }
            creator: { $ref: '#/components/schemas/AuthorUser' }
            activity:
              type: array
              description: |
                El feed del requisito. **El handler la rebautiza:** la asociación es
                `requirementActivities` y la clave de la respuesta es `activity`
                (`requirements-id-get.ts:43-44`). Solo la devuelve
                `GET /requirements/{reqid}`.
              items:
                allOf:
                  - $ref: '#/components/schemas/RequirementActivity'
                  - type: object
                    properties:
                      changedBy: { type: string }
                      previousValue: { type: string }
                      newValue: { type: string }
                      changedByUser: { $ref: '#/components/schemas/AuthorUser' }
            responsiblePeople:
              type: array
              description: |
                Aplanado por el handler: `isLeader` sale de la tabla intermedia
                `people_requirements` y queda al mismo nivel que los campos de la persona
              items:
                type: object
                properties:
                  id: { type: integer }
                  firstName: { type: string }
                  lastName: { type: string }
                  isLeader: { type: boolean, nullable: true }
            totalMinutes:
              type: integer
              description: |
                Minutos imputados al requisito más los de sus tareas. **Solo presente con
                `include=totalMinutes` en `GET /requirements`**: son dos subconsultas
                correlacionadas por fila y no se pagan si no se piden. Misma gradación que el
                incluible homónimo de `core-queries.yaml`. Mismo valor que devuelve
                `GET /requirements/{reqid}/worked-hours`.
    RequirementActivity:
      type: object
      properties:
        id: { type: integer }
        requirementId: { type: integer }
        typeOfActivity: { type: string }
        visibilityLevel: { $ref: '#/components/schemas/VisibilityLevel' }
        createdAt: { type: string, format: date-time }
...